tools: Move codeowners library from cilium-cli dir
During creation of the v1.18 branch, it was problematic to have this
library live in cilium-cli due to other assumptions made in the tree
about the ability to delete this directory while creating a stable
branch. Move this library to another location so we can just delete the
cilium-cli directory.
